refactor(e2e): move message box stub into src/main/e2e
The dialog stub is E2E-only -- its sole consumer is maybeRegisterE2eHooks(),
which no-ops unless NODE_ENV=test -- but it landed in src/main/ root, where
it reads as production code. Move it alongside the rest of that family
(badgeState, notificationClick, trayMenu, hooks). Pure rename plus the
import update; no behaviour change.
