+++ src/main/e2e/messageBoxStub.ts
